Fix CMake build for macOS/Unix: all 24 variants compile successfully
- Update cmake_minimum_required from 2.8 to 3.10
- Add -Dunix to Unix Makefiles definitions for correct platform detection
- Add -std=legacy for gfortran compatibility with Fortran 77 code
- Use static libraries instead of shared on Unix to resolve circular deps
- Add .inc to include directory detection for wdbkwtdata.inc
- Fix include path ordering: bypass include_directories() and use
  CMAKE_<LANG>_FLAGS directly so variant-specific dirs (e.g. ak/common)
  are searched before generic dirs (e.g. common)
- Fix C++ narrowing conversion in fire/cfim/cfim.cpp
- Add missing fiavbc source files to FVSbc and FVSon source lists
  (dbs_fiavbc_atrtls.f, dbs_fiavbc_cutlst.f, dbs_fiavbc_trls.f,
  dbsreference.f, wdbkwtdata.inc)

+++ fire/cfim/cfim.cpp
@@ -528,7 +528,7 @@ double CalcSpreadRate(double *Fuel, double *Moisture, double WindSpeed,
      if(ndead>0)
           ndead=3;
 
-	double nclas[2]={ndead,nlive};  // # of dead & live fuel classes
+	double nclas[2]={(double)ndead,(double)nlive};  // # of dead & live fuel classes
 
    double load[3][2]=			// tons per acre, later converted to lb/ft2
 	{	{Fuel[0], Fuel[3]},
